# Supports serializing a range of rows for UnsafeRowFast