--- a/velox/serializers/UnsafeRowSerializer.cpp
+++ b/velox/serializers/UnsafeRowSerializer.cpp
@@ -34,17 +34,28 @@ class UnsafeRowVectorSerializer : public IterativeVectorSerializer {
       const folly::Range<const IndexRange*>& ranges,
       Scratch& /*scratch*/) override {
     size_t totalSize = 0;
+    const auto totalRows = std::accumulate(
+        ranges.begin(),
+        ranges.end(),
+        0,
+        [](vector_size_t sum, const auto& range) { return sum + range.size; });
+
+    if (totalRows == 0) {
+      return;
+    }
+
     row::UnsafeRowFast unsafeRow(vector);
+    std::vector<vector_size_t> rowSize(totalRows);
     if (auto fixedRowSize =
             row::UnsafeRowFast::fixedRowSize(asRowType(vector->type()))) {
-      for (const auto& range : ranges) {
-        totalSize += (fixedRowSize.value() + sizeof(TRowSize)) * range.size;
-      }
-
+      totalSize += (fixedRowSize.value() + sizeof(TRowSize)) * totalRows;
+      std::fill(rowSize.begin(), rowSize.end(), fixedRowSize.value());
     } else {
+      vector_size_t index = 0;
       for (const auto& range : ranges) {
-        for (auto i = range.begin; i < range.begin + range.size; ++i) {
-          totalSize += unsafeRow.rowSize(i) + sizeof(TRowSize);
+        for (auto i = 0; i < range.size; ++i, ++index) {
+          rowSize[index] = unsafeRow.rowSize(range.begin + i);
+          totalSize += rowSize[index] + sizeof(TRowSize);
         }
       }
     }
@@ -54,19 +65,29 @@ class UnsafeRowVectorSerializer : public IterativeVectorSerializer {
     }
 
     BufferPtr buffer = AlignedBuffer::allocate<char>(totalSize, pool_, 0);
-    auto rawBuffer = buffer->asMutable<char>();
+    auto* rawBuffer = buffer->asMutable<char>();
     buffers_.push_back(std::move(buffer));
 
     size_t offset = 0;
-    for (auto& range : ranges) {
-      for (auto i = range.begin; i < range.begin + range.size; ++i) {
-        // Write row data.
-        TRowSize size =
-            unsafeRow.serialize(i, rawBuffer + offset + sizeof(TRowSize));
-
-        // Write raw size. Needs to be in big endian order.
-        *(TRowSize*)(rawBuffer + offset) = folly::Endian::big(size);
-        offset += sizeof(TRowSize) + size;
+    vector_size_t index = 0;
+    for (const auto& range : ranges) {
+      if (range.size == 1) {
+        // Fast path for single-row serialization.
+        *(TRowSize*)(rawBuffer + offset) = folly::Endian::big(rowSize[index]);
+        auto size = unsafeRow.serialize(
+            range.begin, rawBuffer + offset + sizeof(TRowSize));
+        offset += size + sizeof(TRowSize);
+        ++index;
+      } else {
+        raw_vector<size_t> offsets(range.size);
+        for (auto i = 0; i < range.size; ++i, ++index) {
+          // Write raw size. Needs to be in big endian order.
+          *(TRowSize*)(rawBuffer + offset) = folly::Endian::big(rowSize[index]);
+          offsets[i] = offset + sizeof(TRowSize);
+          offset += rowSize[index] + sizeof(TRowSize);
+        }
+        // Write row data for all rows in range.
+        unsafeRow.serialize(range.begin, range.size, offsets.data(), rawBuffer);
       }
     }
   }