Adopt similar optimization from https://github.com/facebookincubator/velox/pull/10714

Peformance improvment of benchmark `velox_unsafe_row_serialize_benchmark - unsafe_serialize`

```
| Test Name            | Time (us) - before            | Time (us) - after                   | Improvement (%) |
|----------------------|-------------------------------|-------------------------------------|-----------------|
| fixedWidth5          | 43.86                         | 22.33                               | 49.09           |
| fixedWidth10         | 80.24                         | 40.67                               | 49.31           |
| fixedWidth20         | 156.73                        | 66.24                               | 57.73           |
| strings1             | 30.76                         | 25.12                               | 18.33           |
| strings5             | 89.66                         | 77.42                               | 13.65           |
| arrays               | 43.64                         | 36.72                               | 15.86           |
| nestedArrays         | 251.33                        | 247.79                              | 1.41            |
| maps                 | 109.51                        | 105.94                              | 3.26            |
| structs              | 70.81                         | 67.46                               | 4.73            |

```